fix structure to make the cast non-breaking
