+++ example/SessionPoolExample.py
